Create market_scraper.py
MARKET SCRAPER: responsible for fetching up-to-date stock market data from Investing.com (Egypt Equities page). Instead of relying on static or outdated CSVs, the scraper ensures that the chatbot always works with live market trends. It uses Crawl4AI to load the Investing.com equities page in a headless browser (runs invisibly in the background). It parses the HTML with pandas.read_html. And finds the right table by checking key columns, stores it in df, cleans it and converts it to JSON.
This model scrapes EGX30 but we could potentially expand to EGX70, 100 or larger for better recommendations